Refactor PartitionBasedMvRefreshProcessor
