+++ fe/fe-core/src/main/java/com/starrocks/scheduler/BaseTaskRunProcessor.java
@@ -41,7 +41,7 @@ public abstract class BaseTaskRunProcessor implements TaskRunProcessor {
         }
     }
 
-    protected void auditAfterExec(TaskRunContext context, StatementBase parsedStmt, PQueryStatistics statistics) {
+    public void auditAfterExec(TaskRunContext context, StatementBase parsedStmt, PQueryStatistics statistics) {
         String origStmt = context.getDefinition();
         ConnectContext ctx = context.getCtx();
         ConnectProcessor processor = new ConnectProcessor(ctx);
