+++ fe/fe-core/src/test/java/com/starrocks/catalog/DropPartitionWithExprListTest.java
@@ -16,7 +16,7 @@ package com.starrocks.catalog;
 
 import com.google.common.collect.Lists;
 import com.starrocks.clone.DynamicPartitionScheduler;
-import com.starrocks.scheduler.PartitionBasedMvRefreshProcessor;
+import com.starrocks.scheduler.mv.MVPCTBasedRefreshProcessor;
 import com.starrocks.server.GlobalStateMgr;
 import com.starrocks.sql.ast.StatementBase;
 import com.starrocks.sql.optimizer.rule.transformation.materialization.MVTestBase;
@@ -576,7 +576,7 @@ public class DropPartitionWithExprListTest extends MVTestBase {
                                     MaterializedView mv = starRocksAssert.getMv("test", mvName);
                                     {
                                         // all partitions are expired, no need to create partitions for mv
-                                        PartitionBasedMvRefreshProcessor processor = refreshMV("test", mv);
+                                        MVPCTBasedRefreshProcessor processor = refreshMV("test", mv);
                                         Assertions.assertEquals(0, mv.getVisiblePartitions().size());
                                         Assertions.assertTrue(processor.getNextTaskRun() == null);
                                         ExecPlan execPlan = processor.getMvContext().getExecPlan();
@@ -591,7 +591,7 @@ public class DropPartitionWithExprListTest extends MVTestBase {
                                         addListPartition(tableName, "p6", "guangdong",
                                                 now.minusMonths(1).format(DateTimeFormatter.ofPattern("yyyy-MM-dd")), true);
 
-                                        PartitionBasedMvRefreshProcessor processor = refreshMV("test", mv);
+                                        MVPCTBasedRefreshProcessor processor = refreshMV("test", mv);
                                         Assertions.assertTrue(processor != null);
                                         Assertions.assertTrue(processor.getNextTaskRun() == null);
                                         Assertions.assertEquals(2, mv.getVisiblePartitions().size());
@@ -624,7 +624,7 @@ public class DropPartitionWithExprListTest extends MVTestBase {
                                     MaterializedView mv = starRocksAssert.getMv("test", mvName);
                                     {
                                         // all partitions are expired, no need to create partitions for mv
-                                        PartitionBasedMvRefreshProcessor processor = refreshMV("test", mv);
+                                        MVPCTBasedRefreshProcessor processor = refreshMV("test", mv);
                                         Assertions.assertEquals(2, mv.getVisiblePartitions().size());
                                         Assertions.assertTrue(processor.getNextTaskRun() == null);
                                         ExecPlan execPlan = processor.getMvContext().getExecPlan();
@@ -639,7 +639,7 @@ public class DropPartitionWithExprListTest extends MVTestBase {
 
                                     {
                                         // all partitions are expired, no need to create partitions for mv
-                                        PartitionBasedMvRefreshProcessor processor = refreshMV("test", mv);
+                                        MVPCTBasedRefreshProcessor processor = refreshMV("test", mv);
                                         Assertions.assertEquals(2, mv.getVisiblePartitions().size());
                                         Assertions.assertTrue(processor.getNextTaskRun() == null);
                                         ExecPlan execPlan = processor.getMvContext().getExecPlan();
@@ -654,7 +654,7 @@ public class DropPartitionWithExprListTest extends MVTestBase {
                                         addListPartition(tableName, "p6", "guangdong",
                                                 now.minusMonths(1).format(DateTimeFormatter.ofPattern("yyyy-MM-dd")), true);
 
-                                        PartitionBasedMvRefreshProcessor processor = refreshMV("test", mv);
+                                        MVPCTBasedRefreshProcessor processor = refreshMV("test", mv);
                                         Assertions.assertTrue(processor != null);
                                         Assertions.assertTrue(processor.getNextTaskRun() == null);
                                         Assertions.assertEquals(4, mv.getVisiblePartitions().size());
