+++ fe/fe-core/src/test/java/com/starrocks/catalog/MVPartitionExprResolverTest.java
@@ -25,16 +25,14 @@ import com.starrocks.analysis.TableName;
 import com.starrocks.common.util.UUIDUtil;
 import com.starrocks.mv.analyzer.MVPartitionExpr;
 import com.starrocks.mv.analyzer.MVPartitionExprResolver;
-import com.starrocks.qe.StmtExecutor;
-import com.starrocks.scheduler.PartitionBasedMvRefreshProcessor;
 import com.starrocks.scheduler.Task;
 import com.starrocks.scheduler.TaskBuilder;
 import com.starrocks.scheduler.TaskRun;
 import com.starrocks.scheduler.TaskRunBuilder;
+import com.starrocks.scheduler.mv.MVPCTBasedRefreshProcessor;
 import com.starrocks.server.GlobalStateMgr;
 import com.starrocks.sql.ast.QueryStatement;
 import com.starrocks.sql.optimizer.rule.transformation.materialization.MVTestBase;
-import com.starrocks.sql.parser.SqlParser;
 import com.starrocks.sql.plan.ExecPlan;
 import com.starrocks.utframe.UtFrameUtils;
 import org.junit.jupiter.api.Assertions;
@@ -262,7 +260,7 @@ public class MVPartitionExprResolverTest extends MVTestBase {
                     TaskRun taskRun = TaskRunBuilder.newBuilder(task).build();
                     taskRun.initStatus(UUIDUtil.genUUID().toString(), System.currentTimeMillis());
                     taskRun.executeTaskRun();
-                    PartitionBasedMvRefreshProcessor processor = (PartitionBasedMvRefreshProcessor) taskRun.getProcessor();
+                    MVPCTBasedRefreshProcessor processor = getPartitionBasedRefreshProcessor(taskRun);
                     Map<Table, Set<String>> baseTables = getRefTableRefreshedPartitions(processor);
                     Assertions.assertEquals(2, baseTables.size());
                     Assertions.assertEquals(Sets.newHashSet("p20220101"), baseTables.get(
@@ -273,9 +271,7 @@ public class MVPartitionExprResolverTest extends MVTestBase {
                                     .getTable(testDb.getFullName(), "tbl16")));
 
                     // insert new data into tbl16's p20220202 partition
-                    String insertSql = "insert into tbl16 partition(p20220202) values('2022-02-02', 3, 10);";
-                    new StmtExecutor(connectContext, SqlParser.parseSingleStatement(
-                            insertSql, connectContext.getSessionVariable().getSqlMode())).execute();
+                    executeInsertSql("insert into tbl16 partition(p20220202) values('2022-02-02', 3, 10);");
                     taskRun.executeTaskRun();
                     Assertions.assertEquals(Sets.newHashSet("p20220202"),
                             processor.getMVTaskRunExtraMessage().getMvPartitionsToRefresh());
@@ -305,11 +301,9 @@ public class MVPartitionExprResolverTest extends MVTestBase {
                     taskRun.executeTaskRun();
 
                     // insert new data into tbl16's p20220202 partition
-                    String insertSql = "insert into tbl16 partition(p20220202) values('2022-02-02', 3, 10);";
-                    new StmtExecutor(connectContext, SqlParser.parseSingleStatement(
-                            insertSql, connectContext.getSessionVariable().getSqlMode())).execute();
+                    executeInsertSql("insert into tbl16 partition(p20220202) values('2022-02-02', 3, 10);");
                     taskRun.executeTaskRun();
-                    PartitionBasedMvRefreshProcessor processor = (PartitionBasedMvRefreshProcessor) taskRun.getProcessor();
+                    MVPCTBasedRefreshProcessor processor = getPartitionBasedRefreshProcessor(taskRun);
                     Assertions.assertEquals(Sets.newHashSet("p202202_202203"),
                             processor.getMVTaskRunExtraMessage().getMvPartitionsToRefresh());
                     Assertions.assertEquals("{tbl15=[p20220202, p20220201], tbl16=[p20220202, p20220201]}",
@@ -341,8 +335,7 @@ public class MVPartitionExprResolverTest extends MVTestBase {
                     {
                         taskRun.initStatus(UUIDUtil.genUUID().toString(), System.currentTimeMillis());
                         taskRun.executeTaskRun();
-                        PartitionBasedMvRefreshProcessor processor =
-                                (PartitionBasedMvRefreshProcessor) taskRun.getProcessor();
+                        MVPCTBasedRefreshProcessor processor = getPartitionBasedRefreshProcessor(taskRun);
                         Map<Table, Set<String>> baseTables = getRefTableRefreshedPartitions(processor);
                         Assertions.assertEquals(2, baseTables.size());
                         Assertions.assertEquals(Sets.newHashSet("p20220101"),
@@ -356,12 +349,10 @@ public class MVPartitionExprResolverTest extends MVTestBase {
 
                     {
                         // insert new data into tbl16's p20220202 partition
-                        String insertSql = "insert into tbl16 partition(p20220202) values('2022-02-02', 3, 10);";
-                        new StmtExecutor(connectContext, SqlParser.parseSingleStatement(
-                                insertSql, connectContext.getSessionVariable().getSqlMode())).execute();
+                        executeInsertSql("insert into tbl16 partition(p20220202) values('2022-02-02', 3, 10);");
                         taskRun.executeTaskRun();
-                        PartitionBasedMvRefreshProcessor processor =
-                                (PartitionBasedMvRefreshProcessor) taskRun.getProcessor();
+
+                        MVPCTBasedRefreshProcessor processor = getPartitionBasedRefreshProcessor(taskRun);
                         // 1. updated partition of tbl16 is p20220202
                         // 2. date_trunc('month', p20220202) is '2022-02'
                         // 3. tbl15's associated partitions are p20220201 and p20220202
@@ -396,11 +387,11 @@ public class MVPartitionExprResolverTest extends MVTestBase {
                     taskRun.executeTaskRun();
 
                     {
-                        executeInsertSql(connectContext, "insert into tbl1 partition(p2) values('2022-02-02', 3, 10);");
+                        executeInsertSql("insert into tbl1 partition(p2) values('2022-02-02', 3, 10);");
                         taskRun.initStatus(UUIDUtil.genUUID().toString(), System.currentTimeMillis());
                         taskRun.executeTaskRun();
-                        PartitionBasedMvRefreshProcessor processor =
-                                (PartitionBasedMvRefreshProcessor) taskRun.getProcessor();
+
+                        MVPCTBasedRefreshProcessor processor = getPartitionBasedRefreshProcessor(taskRun);
                         Assertions.assertEquals(Sets.newHashSet("p2"),
                                 processor.getMVTaskRunExtraMessage().getMvPartitionsToRefresh());
                         ExecPlan execPlan = processor.getMvContext().getExecPlan();
@@ -410,11 +401,10 @@ public class MVPartitionExprResolverTest extends MVTestBase {
                     }
 
                     {
-                        executeInsertSql(connectContext, "insert into tbl2 partition(p2) values('2022-02-02', 3, 10);");
+                        executeInsertSql("insert into tbl2 partition(p2) values('2022-02-02', 3, 10);");
                         taskRun.initStatus(UUIDUtil.genUUID().toString(), System.currentTimeMillis());
                         taskRun.executeTaskRun();
-                        PartitionBasedMvRefreshProcessor processor =
-                                (PartitionBasedMvRefreshProcessor) taskRun.getProcessor();
+                        MVPCTBasedRefreshProcessor processor = getPartitionBasedRefreshProcessor(taskRun);
                         ExecPlan execPlan = processor.getMvContext().getExecPlan();
                         Assertions.assertTrue(execPlan != null);
                         assertPlanContains(execPlan, "partitions=5/5\n     rollup: tbl1",
@@ -463,14 +453,13 @@ public class MVPartitionExprResolverTest extends MVTestBase {
                     taskRun.initStatus(UUIDUtil.genUUID().toString(), System.currentTimeMillis());
                     taskRun.executeTaskRun();
 
-                    new StmtExecutor(connectContext, SqlParser.parseSingleStatement(
-                            "insert into tbl15 partition(p20220202) values('2022-02-02', 3, 10);",
-                            connectContext.getSessionVariable().getSqlMode())).execute();
+                    executeInsertSql("insert into tbl15 partition(p20220202) values('2022-02-02', 3, 10);");
                     taskRun.executeTaskRun();
-                    PartitionBasedMvRefreshProcessor processor = (PartitionBasedMvRefreshProcessor) taskRun.getProcessor();
-                    Assertions.assertEquals(Sets.newHashSet("p202201_202202", "p202202_202203"),
+
+                    MVPCTBasedRefreshProcessor processor = getPartitionBasedRefreshProcessor(taskRun);
+                    Assertions.assertEquals(Sets.newHashSet("p202202_202203"),
                             processor.getMVTaskRunExtraMessage().getMvPartitionsToRefresh());
-                    Assertions.assertEquals("{tbl15=[p20220103, p20220202, p20220102, p20220201, p20220101]}",
+                    Assertions.assertEquals("{tbl15=[p20220202, p20220201]}",
                             processor.getMVTaskRunExtraMessage().getRefBasePartitionsToRefreshMap().toString());
                 });
     }
@@ -502,9 +491,10 @@ public class MVPartitionExprResolverTest extends MVTestBase {
                     taskRun.initStatus(UUIDUtil.genUUID().toString(), System.currentTimeMillis());
                     taskRun.executeTaskRun();
 
-                    executeInsertSql(connectContext, "insert into tbl2 partition(p1) values('2022-01-02', 3, 10);");
+                    executeInsertSql("insert into tbl2 partition(p1) values('2022-01-02', 3, 10);");
                     taskRun.executeTaskRun();
-                    PartitionBasedMvRefreshProcessor processor = (PartitionBasedMvRefreshProcessor) taskRun.getProcessor();
+
+                    MVPCTBasedRefreshProcessor processor = getPartitionBasedRefreshProcessor(taskRun);
                     Set<String> mvPartitionsToRefresh = processor.getMVTaskRunExtraMessage().getMvPartitionsToRefresh();
                     System.out.println(mvPartitionsToRefresh);
                     Assertions.assertTrue(mvPartitionsToRefresh.contains("p20220101_20220102"));
@@ -616,17 +606,24 @@ public class MVPartitionExprResolverTest extends MVTestBase {
                             ((MaterializedView) GlobalStateMgr.getCurrentState().getLocalMetastore()
                                     .getTable(testDb.getFullName(), "mv_join_predicate"));
                     Task task = TaskBuilder.buildMvTask(materializedView, testDb.getFullName());
-                    TaskRun taskRun = TaskRunBuilder.newBuilder(task).build();
-                    taskRun.initStatus(UUIDUtil.genUUID().toString(), System.currentTimeMillis());
-                    taskRun.executeTaskRun();
+                    {
+                        TaskRun taskRun = TaskRunBuilder.newBuilder(task).build();
+                        taskRun.initStatus(UUIDUtil.genUUID().toString(), System.currentTimeMillis());
+                        taskRun.executeTaskRun();
+                    }
 
-                    executeInsertSql(connectContext, "insert into tbl2 partition(p1) values('2022-01-02', 3, 10);");
-                    taskRun.executeTaskRun();
-                    PartitionBasedMvRefreshProcessor processor = (PartitionBasedMvRefreshProcessor) taskRun.getProcessor();
-                    Assertions.assertEquals(Sets.newHashSet("p1"),
-                            processor.getMVTaskRunExtraMessage().getMvPartitionsToRefresh());
-                    Assertions.assertEquals("{tbl2=[p1], tbl1=[p1]}",
-                            processor.getMVTaskRunExtraMessage().getRefBasePartitionsToRefreshMap().toString());
+                    {
+                        TaskRun taskRun = TaskRunBuilder.newBuilder(task).build();
+                        executeInsertSql("insert into tbl2 partition(p1) values('2022-01-02', 3, 10);");
+                        taskRun.initStatus(UUIDUtil.genUUID().toString(), System.currentTimeMillis());
+                        taskRun.executeTaskRun();
+
+                        MVPCTBasedRefreshProcessor processor = getPartitionBasedRefreshProcessor(taskRun);
+                        Assertions.assertEquals(Sets.newHashSet("p1"),
+                                processor.getMVTaskRunExtraMessage().getMvPartitionsToRefresh());
+                        Assertions.assertEquals("{tbl2=[p1], tbl1=[p1]}",
+                                processor.getMVTaskRunExtraMessage().getRefBasePartitionsToRefreshMap().toString());
+                    }
                 });
     }
 }
