+++ fe/fe-core/src/test/java/com/starrocks/sql/optimizer/rule/transformation/materialization/MVTestBase.java
@@ -38,13 +38,15 @@ import com.starrocks.pseudocluster.PseudoCluster;
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.qe.StmtExecutor;
 import com.starrocks.scheduler.ExecuteOption;
+import com.starrocks.scheduler.MVTaskRunProcessor;
 import com.starrocks.scheduler.MvTaskRunContext;
-import com.starrocks.scheduler.PartitionBasedMvRefreshProcessor;
-import com.starrocks.scheduler.TableSnapshotInfo;
 import com.starrocks.scheduler.Task;
 import com.starrocks.scheduler.TaskBuilder;
 import com.starrocks.scheduler.TaskRun;
 import com.starrocks.scheduler.TaskRunBuilder;
+import com.starrocks.scheduler.TaskRunProcessor;
+import com.starrocks.scheduler.mv.BaseTableSnapshotInfo;
+import com.starrocks.scheduler.mv.MVPCTBasedRefreshProcessor;
 import com.starrocks.server.GlobalStateMgr;
 import com.starrocks.sql.analyzer.Analyzer;
 import com.starrocks.sql.ast.CreateMaterializedViewStatement;
@@ -94,7 +96,20 @@ import java.util.stream.Collectors;
 /**
  * Base class for materialized view tests.
  */
-public class MVTestBase extends StarRocksTestBase {
+public abstract class MVTestBase extends StarRocksTestBase {
+
+    public interface ExceptionRunnable {
+        void run() throws Exception;
+    }
+
+    public interface ExecPlanChecker {
+        void check(ExecPlan execPlan) throws Exception;
+    }
+
+    public interface MVActionRunner {
+        void run(MaterializedView mv) throws Exception;
+    }
+
     protected static final Logger LOG = LogManager.getLogger(MVTestBase.class);
     protected static ConnectContext connectContext;
     protected static PseudoCluster cluster;
@@ -353,14 +368,24 @@ public class MVTestBase extends StarRocksTestBase {
         refreshMVRange(mvName, null, null, force);
     }
 
-    protected PartitionBasedMvRefreshProcessor refreshMV(String dbName, MaterializedView mv) throws Exception {
+    protected TaskRun withMVRefreshTaskRun(String dbName, MaterializedView mv) throws Exception {
         Task task = TaskBuilder.buildMvTask(mv, dbName);
         Map<String, String> testProperties = task.getProperties();
         testProperties.put(TaskRun.IS_TEST, "true");
         TaskRun taskRun = TaskRunBuilder.newBuilder(task).build();
         taskRun.initStatus(UUIDUtil.genUUID().toString(), System.currentTimeMillis());
         taskRun.executeTaskRun();
-        return (PartitionBasedMvRefreshProcessor) taskRun.getProcessor();
+        return taskRun;
+    }
+
+    protected MVTaskRunProcessor withMVRefreshProcessor(String dbName, MaterializedView mv) throws Exception {
+        TaskRun taskRun = withMVRefreshTaskRun(dbName, mv);
+        return getMVRefreshProcessor(taskRun);
+    }
+
+    protected MVPCTBasedRefreshProcessor refreshMV(String dbName, MaterializedView mv) throws Exception {
+        TaskRun taskRun = withMVRefreshTaskRun(dbName, mv);
+        return getPartitionBasedRefreshProcessor(taskRun);
     }
 
     protected void refreshMVRange(String mvName, String start, String end, boolean force) throws Exception {
@@ -385,9 +410,9 @@ public class MVTestBase extends StarRocksTestBase {
                 QueryMaterializationContext.QueryCacheStats.class);
     }
 
-    protected Map<Table, Set<String>> getRefTableRefreshedPartitions(PartitionBasedMvRefreshProcessor processor) {
-        Map<TableSnapshotInfo, Set<String>> baseTables = processor
-                .getRefTableRefreshPartitions(Sets.newHashSet("p20220101"));
+    protected Map<Table, Set<String>> getRefTableRefreshedPartitions(MVPCTBasedRefreshProcessor processor) {
+        Map<BaseTableSnapshotInfo, Set<String>> baseTables = processor
+                .getPCTRefTableRefreshPartitions(Sets.newHashSet("p20220101"));
         Assertions.assertEquals(2, baseTables.size());
         return baseTables.entrySet().stream().collect(Collectors.toMap(x -> x.getKey().getBaseTable(), x -> x.getValue()));
     }
@@ -403,9 +428,10 @@ public class MVTestBase extends StarRocksTestBase {
 
     protected static ExecPlan getMVRefreshExecPlan(TaskRun taskRun) throws Exception {
         initAndExecuteTaskRun(taskRun);
-        PartitionBasedMvRefreshProcessor processor = (PartitionBasedMvRefreshProcessor)
-                taskRun.getProcessor();
-        MvTaskRunContext mvTaskRunContext = processor.getMvContext();
+        TaskRunProcessor processor = taskRun.getProcessor();
+        Assertions.assertTrue(processor instanceof MVTaskRunProcessor);
+        MVTaskRunProcessor mvTaskRunProcessor = (MVTaskRunProcessor) processor;
+        MvTaskRunContext mvTaskRunContext = mvTaskRunProcessor.getMvTaskRunContext();
         return mvTaskRunContext.getExecPlan();
     }
 
@@ -553,9 +579,7 @@ public class MVTestBase extends StarRocksTestBase {
         }
     }
 
-    public interface ExceptionRunnable {
-        public abstract void run() throws Exception;
-    }
+
 
     protected void doTest(List<TestListener> listeners, ExceptionRunnable testCase) {
         for (TestListener listener : listeners) {
@@ -591,8 +615,7 @@ public class MVTestBase extends StarRocksTestBase {
             Task task = TaskBuilder.buildMvTask(mv, testDb.getFullName());
             TaskRun taskRun = TaskRunBuilder.newBuilder(task).setExecuteOption(executeOption).build();
             initAndExecuteTaskRun(taskRun);
-            PartitionBasedMvRefreshProcessor processor =
-                    (PartitionBasedMvRefreshProcessor) taskRun.getProcessor();
+            MVPCTBasedRefreshProcessor processor = getPartitionBasedRefreshProcessor(taskRun);
             MvTaskRunContext mvTaskRunContext = processor.getMvContext();
             ExecPlan execPlan = mvTaskRunContext.getExecPlan();
             Assertions.assertTrue(execPlan != null);
@@ -600,5 +623,29 @@ public class MVTestBase extends StarRocksTestBase {
                 assertPlanContains(execPlan, expectStr);
             }
         });
+    }
+
+    public static MVTaskRunProcessor getMVRefreshProcessor(TaskRun taskRun) {
+        Assertions.assertTrue(taskRun.getProcessor() instanceof MVTaskRunProcessor);
+        return (MVTaskRunProcessor) taskRun.getProcessor();
+    }
+
+    public static MVPCTBasedRefreshProcessor getPartitionBasedRefreshProcessor(TaskRun taskRun) {
+        Assertions.assertTrue(taskRun.getProcessor() instanceof MVTaskRunProcessor);
+        MVTaskRunProcessor mvTaskRunProcessor = (MVTaskRunProcessor) taskRun.getProcessor();
+        return (MVPCTBasedRefreshProcessor) mvTaskRunProcessor.getMVRefreshProcessor();
+    }
+
+    protected void withMVQuery(String mvQuery,
+                               MVActionRunner runner) throws Exception {
+        String ddl = String.format("CREATE MATERIALIZED VIEW `test`.`test_mv1` " +
+                "REFRESH DEFERRED MANUAL\n" +
+                "PROPERTIES (\n" +
+                "\"refresh_mode\" = \"incremental\"" +
+                ")\n" +
+                "AS %s;", mvQuery);
+        starRocksAssert.withMaterializedView(ddl);
+        MaterializedView mv = getMv("test_mv1");
+        runner.run(mv);
     }
 }
